# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bureau Agent API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [en_attente, en_cours, termine, annule, all]
        - name: priority
          in: query
          schema:
            type: string
            enum: [haute, moyenne, basse, all]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: "#/components/schemas/Task"
                  total:
                    type: integer
                required: [tasks, total]
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # === MESSAGES ===
  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List voicemail/messages
      parameters:
        - name: read
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: "#/components/schemas/Message"
                  total:
                    type: integer
                required: [messages, total]
    post:
      operationId: createMessage
      tags: [messages]
      summary: Create a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageBody"
      responses:
        "201":
          description: Message created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /messages/{id}:
    get:
      operationId: getMessage
      tags: [messages]
      summary: Get a message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Message detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    patch:
      operationId: updateMessage
      tags: [messages]
      summary: Update a message (mark as read, etc.)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMessageBody"
      responses:
        "200":
          description: Updated message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # === DASHBOARD / ANALYTICS ===
...
  /dashboard/call-analytics:
    get:
      operationId: getCallAnalytics
      tags: [dashboard]
      summary: Call analytics over time
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, week, month, year]
            default: week
      responses:
        "200":
          description: Call analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallAnalytics"
  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Recent activity feed
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Recent activity items
          content:
            application/json:
              schema:
                type: object
                properties:
                  activities:
                    type: array
                    items:
                      $ref: "#/components/schemas/ActivityItem"
                required: [activities]
...
  /dashboard/top-contacts:
    get:
      operationId: getTopContacts
      tags: [dashboard]
      summary: Most active contacts
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Top contacts
          content:
            application/json:
              schema:
                type: object
                properties:
                  contacts:
                    type: array
                    items:
                      $ref: "#/components/schemas/TopContact"
                required: [contacts]
components:
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [en_attente, en_cours, termine, annule]
        priority:
          type: string
          enum: [haute, moyenne, basse]
        dueDate:
          type: string
          format: date-time
          nullable: true
        assignedTo:
          type: string
          nullable: true
        relatedContactId:
          type: integer
          nullable: true
        relatedCallId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, status, priority, createdAt, updatedAt]
    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [en_attente, en_cours, termine, annule]
        priority:
          type: string
          enum: [haute, moyenne, basse]
        dueDate:
          type: string
          format: date-time
          nullable: true
        assignedTo:
          type: string
          nullable: true
        relatedContactId:
          type: integer
          nullable: true
        relatedCallId:
          type: integer
          nullable: true
      required: [title, status, priority]
    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [en_attente, en_cours, termine, annule]
        priority:
          type: string
          enum: [haute, moyenne, basse]
        dueDate:
          type: string
          format: date-time
          nullable: true
        assignedTo:
          type: string
          nullable: true
        relatedContactId:
          type: integer
          nullable: true
        relatedCallId:
          type: integer
          nullable: true
    Message:
      type: object
      properties:
        id:
          type: integer
        contactId:
          type: integer
          nullable: true
        contactName:
          type: string
          nullable: true
        phoneNumber:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [messagerie_vocale, note, rappel]
        isRead:
          type: boolean
        priority:
          type: string
          enum: [haute, moyenne, basse]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, phoneNumber, content, type, isRead, priority, createdAt, updatedAt]
    CreateMessageBody:
      type: object
      properties:
        contactId:
          type: integer
          nullable: true
        phoneNumber:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [messagerie_vocale, note, rappel]
        priority:
          type: string
          enum: [haute, moyenne, basse]
      required: [phoneNumber, content, type, priority]
    UpdateMessageBody:
      type: object
      properties:
        isRead:
          type: boolean
        content:
          type: string
        priority:
          type: string
          enum: [haute, moyenne, basse]
...
    CallAnalytics:
      type: object
      properties:
        period:
          type: string
        dataPoints:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              answered:
                type: integer
              missed:
                type: integer
              voicemail:
                type: integer
              avgDuration:
                type: number
            required: [label, answered, missed, voicemail, avgDuration]
        totalAnswered:
          type: integer
        totalMissed:
          type: integer
        totalVoicemail:
          type: integer
        answerRate:
          type: number
      required: [period, dataPoints, totalAnswered, totalMissed, totalVoicemail, answerRate]
    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [appel, contact, tache, message]
        description:
          type: string
        timestamp:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
      required: [id, type, description, timestamp]
...
    TopContact:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        company:
          type: string
          nullable: true
        totalCalls:
          type: integer
        lastCallAt:
          type: string
          format: date-time
          nullable: true
      required: [id, firstName, lastName, totalCalls]